Refactored navbar by splitting navbar.jsx into multiple components
