+++ frontend/src/components/stylesheets/LogoutButton.css
@@ -1,0 +1,19 @@
+.logout-button {
+    width: 100%;
+    text-align: left;
+    background: none;
+    border: none;
+    border-radius: var(--radius-md);
+    padding: var(--space-sm) var(--space-md);
+    color: var(--danger-soft);
+    font-size: var(--font-size-sm);
+    display: flex;
+    align-items: center;
+    gap: var(--space-sm);
+    box-shadow: none;
+    cursor: pointer;
+}
+
+.logout-button:hover {
+    background-color: var(--bg-soft-hover);
+}
