--- frontend/src/components/stylesheets/Navbar.css
@@ -1,193 +1,0 @@
-.navbar-container {
-    justify-content: space-between;
-    display: flex;
-    align-items: center;
-    padding: var(--space-lg) var(--space-navbar-x);
-    margin-bottom: var(--space-lg);
-    border-bottom: 6px solid var(--border-light);
-    background: var(--bg-main);
-}
-
-.navbar-left {
-    display: flex;
-    align-items: center;
-    gap: var(--space-lg);
-}
-
-.navbar-title {
-    text-decoration: none;
-    color: var(--text-main);
-    font-family: var(--font-main), sans-serif;
-    font-size: var(--font-size-xl);
-    font-weight: bold;
-}
-
-.navbar-title:hover {
-    color: var(--text-dark);
-}
-
-.navbar-link {
-    display: flex;
-    align-items: center;
-    gap: var(--space-xs);
-    text-decoration: none;
-    color: var(--text-muted);
-    font-family: var(--font-main), sans-serif;
-    font-size: 1.1rem;
-    font-weight: bold;
-    line-height: 1;
-}
-
-.navbar-link:hover {
-    color: var(--text-main);
-}
-
-.navbar-icon {
-    font-size: 20px;
-    flex-shrink: 0;
-    opacity: 0.6;
-}
-
-.navbar-right {
-    display: flex;
-    align-items: center;
-    gap: var(--space-md);
-}
-
-.navbar-tagline {
-    font-size: var(--font-size-sm);
-    color: var(--text-muted);
-}
-
-.navbar-user {
-    position: relative;
-}
-
-.user-icon {
-    font-size: 38px;
-    cursor: pointer;
-    color: var(--text-muted);
-    opacity: 0.6;
-    transition:
-        color 0.2s,
-        opacity 0.2s;
-}
-
-.user-icon:hover {
-    opacity: 1;
-    color: var(--text-main);
-}
-
-.navbar-dropdown {
-    position: fixed;
-    top: 75px;
-    right: var(--space-navbar-x);
-    min-width: 180px;
-    background: var(--bg-soft);
-    color: var(--text-main);
-    border: 1px solid var(--border-light);
-    box-shadow: var(--shadow-md);
-    border-radius: var(--radius-md);
-    padding: var(--space-xs);
-    z-index: 1000;
-    display: flex;
-    flex-direction: column;
-    gap: var(--space-xs);
-}
-
-.dropdown-username {
-    font-size: var(--font-size-sm);
-    color: var(--text-dark);
-    font-weight: normal;
-    padding: var(--space-sm) var(--space-md);
-}
-
-.navbar-dropdown-divider {
-    border: none;
-    border-top: 1px solid var(--border-strong);
-    margin: 0;
-}
-
-.dropdown-link {
-    text-decoration: none;
-    color: var(--text-main);
-    font-size: var(--font-size-sm);
-    padding: var(--space-sm) var(--space-md);
-    border-radius: var(--radius-md);
-    background: none;
-    display: flex;
-    align-items: center;
-    gap: var(--space-sm);
-}
-
-.dropdown-link:hover {
-    background-color: var(--bg-soft-hover);
-}
-
-.theme-toggle-btn {
-    background: var(--bg-soft);
-    color: var(--text-main);
-    border: 1px solid var(--border-light);
-    border-radius: var(--radius-md);
-    margin-right: var(--space-md);
-}
-
-.logout-button {
-    width: 100%;
-    text-align: left;
-    background: none;
-    border: none;
-    border-radius: var(--radius-md);
-    padding: var(--space-sm) var(--space-md);
-    color: var(--danger-soft);
-    font-size: var(--font-size-sm);
-    display: flex;
-    align-items: center;
-    gap: var(--space-sm);
-    box-shadow: none;
-    cursor: pointer;
-}
-
-.logout-button:hover {
-    background-color: var(--bg-soft-hover);
-}
-
-@media (max-width: 768px) {
-    .navbar-container {
-        padding: var(--space-sm) var(--space-md);
-    }
-
-    .navbar-title {
-        font-size: var(--font-size-lg);
-    }
-
-    .navbar-link span {
-        display: none;
-    }
-
-    .user-dropdown {
-        min-width: 140px;
-    }
-}
-
-@media (max-width: 360px) {
-    .navbar-container {
-        padding: var(--space-xs) var(--space-sm);
-    }
-
-    .navbar-left {
-        gap: var(--space-md);
-    }
-
-    .navbar-right {
-        gap: var(--space-xs);
-    }
-
-    .navbar-title {
-        font-size: 1.2rem;
-    }
-
-    .user-icon {
-        font-size: 32px;
-    }
-}
