+++ frontend/src/components/stylesheets/Navbar/NavbarBrand.css
@@ -1,0 +1,23 @@
+.navbar-title {
+    text-decoration: none;
+    color: var(--text-main);
+    font-family: var(--font-main), sans-serif;
+    font-size: var(--font-size-xl);
+    font-weight: bold;
+}
+
+.navbar-title:hover {
+    color: var(--text-dark);
+}
+
+@media (max-width: 768px) {
+    .navbar-title {
+        font-size: var(--font-size-lg);
+    }
+}
+
+@media (max-width: 360px) {
+    .navbar-title {
+        font-size: var(--font-size-message);
+    }
+}
