+++ frontend/src/components/stylesheets/Navbar/NavbarDropdownMenu.css
@@ -1,0 +1,45 @@
+.navbar-dropdown {
+    position: fixed;
+    top: var(--navbar-height);
+    right: var(--space-navbar-x);
+    min-width: 180px;
+    background: var(--bg-soft);
+    color: var(--text-main);
+    border: 1px solid var(--border-light);
+    box-shadow: var(--shadow-md);
+    border-radius: var(--radius-md);
+    padding: var(--space-xs);
+    z-index: var(--z-index-popup);
+    display: flex;
+    flex-direction: column;
+    gap: var(--space-xs);
+}
+
+.dropdown-username {
+    font-size: var(--font-size-sm);
+    color: var(--text-dark);
+    font-weight: normal;
+    padding: var(--space-sm) var(--space-md);
+}
+
+.navbar-dropdown-divider {
+    border: none;
+    border-top: 1px solid var(--border-strong);
+    margin: 0;
+}
+
+.dropdown-link {
+    text-decoration: none;
+    color: var(--text-main);
+    font-size: var(--font-size-sm);
+    padding: var(--space-sm) var(--space-md);
+    border-radius: var(--radius-md);
+    background: none;
+    display: flex;
+    align-items: center;
+    gap: var(--space-sm);
+}
+
+.dropdown-link:hover {
+    background-color: var(--bg-soft-hover);
+}
