+++ frontend/src/components/stylesheets/Navbar/NavbarLinks.css
@@ -1,0 +1,27 @@
+.navbar-link {
+    display: flex;
+    align-items: center;
+    gap: var(--space-xs);
+    text-decoration: none;
+    color: var(--text-muted);
+    font-family: var(--font-main), sans-serif;
+    font-size: var(--font-size-base);
+    font-weight: bold;
+    line-height: 1;
+}
+
+.navbar-link:hover {
+    color: var(--text-main);
+}
+
+.navbar-icon {
+    font-size: var(--font-size-heading);
+    flex-shrink: 0;
+    opacity: 0.6;
+}
+
+@media (max-width: 768px) {
+    .navbar-link span {
+        display: none;
+    }
+}
