+++ frontend/src/components/stylesheets/Navbar/NavbarUserDropdown.css
@@ -1,0 +1,24 @@
+.navbar-user {
+    position: relative;
+}
+
+.user-icon {
+    font-size: var(--icon-size-user);
+    cursor: pointer;
+    color: var(--text-muted);
+    opacity: 0.6;
+    transition:
+            color var(--transition-speed),
+            opacity var(--transition-speed);
+}
+
+.user-icon:hover {
+    opacity: 1;
+    color: var(--text-main);
+}
+
+@media (max-width: 360px) {
+    .user-icon {
+        font-size: var(--font-size-lg);
+    }
+}
